Rename `CmdCtx*Builder` to `CmdCtx*Params`.
