+++ crate/cmd_ctx/src/lib.rs
@@ -2,12 +2,19 @@
 //! framework.
 
 pub use crate::{
-    cmd_ctx_mpnf::CmdCtxMpnf, cmd_ctx_mpnf_builder::CmdCtxMpnfBuilder, cmd_ctx_mpsf::CmdCtxMpsf,
-    cmd_ctx_mpsf_builder::CmdCtxMpsfBuilder, cmd_ctx_npnf::CmdCtxNpnf,
-    cmd_ctx_npnf_builder::CmdCtxNpnfBuilder, cmd_ctx_spnf::CmdCtxSpnf,
-    cmd_ctx_spnf_builder::CmdCtxSpnfBuilder, cmd_ctx_spsf::CmdCtxSpsf,
-    cmd_ctx_spsf_builder::CmdCtxSpsfBuilder, cmd_ctx_types::CmdCtxTypes,
-    profile_filter_fn::ProfileFilterFn, profile_selection::ProfileSelection,
+    cmd_ctx_mpnf::CmdCtxMpnf,
+    cmd_ctx_mpnf_params::{CmdCtxMpnfParams, CmdCtxMpnfParamsBuilder},
+    cmd_ctx_mpsf::CmdCtxMpsf,
+    cmd_ctx_mpsf_params::{CmdCtxMpsfParams, CmdCtxMpsfParamsBuilder},
+    cmd_ctx_npnf::CmdCtxNpnf,
+    cmd_ctx_npnf_params::{CmdCtxNpnfParams, CmdCtxNpnfParamsBuilder},
+    cmd_ctx_spnf::CmdCtxSpnf,
+    cmd_ctx_spnf_params::{CmdCtxSpnfParams, CmdCtxSpnfParamsBuilder},
+    cmd_ctx_spsf::CmdCtxSpsf,
+    cmd_ctx_spsf_params::{CmdCtxSpsfParams, CmdCtxSpsfParamsBuilder},
+    cmd_ctx_types::CmdCtxTypes,
+    profile_filter_fn::ProfileFilterFn,
+    profile_selection::ProfileSelection,
 };
 
 pub(crate) use crate::{
@@ -18,15 +25,15 @@ pub(crate) use crate::{
 mod cmd_ctx_builder_support;
 mod cmd_ctx_builder_support_multi;
 mod cmd_ctx_mpnf;
-mod cmd_ctx_mpnf_builder;
+mod cmd_ctx_mpnf_params;
 mod cmd_ctx_mpsf;
-mod cmd_ctx_mpsf_builder;
+mod cmd_ctx_mpsf_params;
 mod cmd_ctx_npnf;
-mod cmd_ctx_npnf_builder;
+mod cmd_ctx_npnf_params;
 mod cmd_ctx_spnf;
-mod cmd_ctx_spnf_builder;
+mod cmd_ctx_spnf_params;
 mod cmd_ctx_spsf;
-mod cmd_ctx_spsf_builder;
+mod cmd_ctx_spsf_params;
 mod cmd_ctx_types;
 mod profile_filter_fn;
 mod profile_selection;
